server_test/mybackend/requirements.txt
asgiref==3.5.2
bidict==0.22.0
Django==4.1.4
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
